--- src/main/kotlin/racingcar/MovingStrategy.kt
@@ -1,5 +1,0 @@
-package racingcar
-
-interface MovingStrategy {
-    fun createNumber() : Int
-}
