+++ src/main/kotlin/racingcar/domain/Car.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 data class Car(val carName: String, private val movingDistance: Int) {
     fun movingDistance(): Int {
