+++ src/main/kotlin/racingcar/domain/MovingStrategy.kt
@@ -1,0 +1,5 @@
+package racingcar.domain
+
+interface MovingStrategy {
+    fun createNumber(): Int
+}
