--- src/main/kotlin/racingcar/ui/InputView.kt
@@ -1,18 +1,0 @@
-package racingcar.ui
-
-class InputView {
-
-    companion object {
-        fun insertCarNames(): List<String> {
-            println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).")
-            return readlnOrNull()?.split(",") ?: emptyList()
-        }
-
-        fun insertRunCount(): Int {
-            println("시도할 횟수는 몇 회인가요?")
-            return readLine()?.toInt() ?: 0
-        }
-    }
-
-
-}
